put save function in base class
